second commit

+++ src/main/java/tn/esprit/gestionachat/Repo/OperateurRepo.java
@@ -1,7 +1,7 @@
 package tn.esprit.gestionachat.Repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import tn.esprit.gestionachat.entities.Operateur;
+import tn.esprit.gestionachat.entity.Operateur;
 
-public interface OperateurRepo extends JpaRepository<Operateur,Integer> {
+public interface OperateurRepo extends JpaRepository<Operateur,Long> {
 }
