+++ src/main/java/tn/esprit/gestionachat/Repo/ProduitRepo.java
@@ -1,7 +1,10 @@
 package tn.esprit.gestionachat.Repo;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-import tn.esprit.gestionachat.entities.Produit;
 
-public interface ProduitRepo extends JpaRepository<Produit,Integer> {
+import org.springframework.data.jpa.repository.JpaRepository;
+import tn.esprit.gestionachat.entity.Produit;
+
+
+public interface ProduitRepo extends JpaRepository<Produit,Long> {
+    //Produit save(Produit p,Long idCategorieProduit, Long idStock);
 }
