+++ src/main/java/tn/esprit/gestionachat/Repo/ReglementRepo.java
@@ -1,7 +1,8 @@
 package tn.esprit.gestionachat.Repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import tn.esprit.gestionachat.entities.Reglement;
+import tn.esprit.gestionachat.entity.Reglement;
 
-public interface ReglementRepo extends JpaRepository<Reglement,Integer> {
+
+public interface ReglementRepo extends JpaRepository<Reglement,Long> {
 }
